fix: remote-azblob-sasToken-Authorization (#1800)
### Description

<!--Add a description of your PR here-->
In case SAS tokens are generated in the level of a container, uploading
the outputs to Azure blob is going to fail on the creation of container.
This fix will pass the authorization exception for such cases. This
fixes #1794

### QC
<!-- Make sure that you can tick the boxes below. -->

- [x] The PR contains a test case for the changes or the changes are
already covered by an existing test case.

* [x] The documentation (`docs/`) is updated to reflect the changes or
this is not necessary (e.g. if the change does neither modify the
language nor the behavior or functionalities of Snakemake).

---------

Co-authored-by: Johannes Köster <johannes.koester@tu-dortmund.de>
Co-authored-by: Johannes Köster <johannes.koester@uni-due.de>