docs: basics.rst: suggest VS Code instead of deprecated Atom as IDE (#2368)
### Description

Atom has been deprecated for quite a while, so we should suggest another
editor / IDE. We have been using VS Code quite a bit and are very
satisfied with its capabilities for developing Snakemake workflows, so
that I think this is a reasonable suggestion. But other suggestions are
welcome.

### QC
<!-- Make sure that you can tick the boxes below. -->

* [ ] The PR contains a test case for the changes or the changes are
already covered by an existing test case.
* [x] The documentation (`docs/`) is updated to reflect the changes or
this is not necessary (e.g. if the change does neither modify the
language nor the behavior or functionalities of Snakemake).

---------

Co-authored-by: Johannes Köster <johannes.koester@uni-due.de>