fix: add missing spaces between lines that get concatenated. (#2268)
### Description

<!--Add a description of your PR here-->
When outputting the help contents, some arguments didn't have spaces
before of after the lines. This meant that for `--cluster-sync`, we
would see `usedif` since the end of the line
`"status upon remote termination (for example, this should be used"` is
joined with `"if the cluster command is 'qsub -sync y' (SGE)"`.

This is a quick PR to change this. Since there were some arguments that
had the spaces before the line and some after, I only changed by adding
some after when missing.

### QC
<!-- Make sure that you can tick the boxes below. -->

* [ ] The PR contains a test case for the changes or the changes are
already covered by an existing test case.
* [ ] The documentation (`docs/`) is updated to reflect the changes or
this is not necessary (e.g. if the change does neither modify the
language nor the behavior or functionalities of Snakemake).

Co-authored-by: Johannes Köster <johannes.koester@tu-dortmund.de>