fix: updating of non-dict config values gives error (#2364)
### Description

This PR addresses an issue in config merging: If a config value that is
not a `dict` (e.g. it is `None`) is replaced by a `dict` an exception is
raised, as `update_config()` tries to update the keys of the config,
which is not possible (e.g. `None` has no keys). This PR fixes this
problem by just overwriting the original value in this case (e.g. `None`
is replaced by the new dictionary).

### QC
<!-- Make sure that you can tick the boxes below. -->

* [x] The PR contains a test case for the changes or the changes are
already covered by an existing test case.
* [x] The documentation (`docs/`) is updated to reflect the changes or
this is not necessary (e.g. if the change does neither modify the
language nor the behavior or functionalities of Snakemake).

---------

Co-authored-by: Johannes Köster <johannes.koester@tu-dortmund.de>
Co-authored-by: Johannes Köster <johannes.koester@uni-due.de>