fix: unnecessary set Snakefile in AzBatch executor (#2397)
### Description

- Removes unnecessary setting of Snakefile in AzBatch executor. This was
1. not necessary, and 2. was causing a failure to resolve the Snakefile
path when it was located at `workflow/Snakefile`.
- Since default resources is now included in the command per #2395 , I
needed to update the shell quoting so the command doesn't fail because
of the parentheses in the default resources. Used the same solution as
#2043 and it works nicely.
- Updated the masking regex to be more generic for SAS token
possibilities

### QC

<!-- Make sure that you can tick the boxes below. -->

* [x] The PR contains a test case for the changes or the changes are
already covered by an existing test case.
* [x] The documentation (`docs/`) is updated to reflect the changes or
this is not necessary (e.g. if the change does neither modify the
language nor the behavior or functionalities of Snakemake).