feat: Faster ci test setup  (#2489)
### Description

This PR speeds up the CI tests.



| Before: | After |
| --- | ----| 
|
![image](https://github.com/snakemake/snakemake/assets/1038978/4e3cf3a3-1a76-472b-8c15-7a3af68fd680)
|
![image](https://github.com/snakemake/snakemake/assets/1038978/418400da-1778-4a25-86df-ab3b25c2fe75)
|

This was done by using
[setup-micromamba](https://github.com/mamba-org/setup-micromamba) to
install mamba and the Python environment. This is especially fast
because the environment is cached.

Here are the speed improvements for micromamba install + environment
creation:

| Platform |  Before: | After: |
| --- | ----| ---- | 
| Linux | 3 min | 11 sec |
| Win | 12 min | 2.5 min | 

(maybe snakemake should use micromamba for its conda-jobs)...

The black formatting job is also much faster: Total time: 60 sec -> 11
sec.

A small speed up also comes from replacing the `cancel-previous` job
with a github-action built-feature (`concurrency`) which does the same.

Finally, a lot of speedup comes from adding a `.test_duration` file so
`pytest-split` can minimize test time by spreading the tests across jobs
in an optimal way. This cuts 30% of the actual test time.


### QC
<!-- Make sure that you can tick the boxes below. -->

* [x] The PR contains a test case for the changes or the changes are
already covered by an existing test case.
* [x] The documentation (`docs/`) is updated to reflect the changes or
this is not necessary (e.g. if the change does neither modify the
language nor the behavior or functionalities of Snakemake).

---------

Co-authored-by: Johannes Köster <johannes.koester@tu-dortmund.de>
Co-authored-by: Johannes Koester <johannes.koester@uni-due.de>