docs: document name directive with example (#2534)
### Description

The `name` directive (#811) is super helpful but I felt it wasn't very
obvious in the docs. I added an example usage based on my own usage of
it as I think a lot of others use it in the same manner.

### QC
<!-- Make sure that you can tick the boxes below. -->

* [ ] The PR contains a test case for the changes or the changes are
already covered by an existing test case.
* [x] The documentation (`docs/`) is updated to reflect the changes or
this is not necessary (e.g. if the change does neither modify the
language nor the behavior or functionalities of Snakemake).