feat: add --sdm short opt for --deployment (#2551)
### Description

The new
`--software-deployment-method`/`--deployment-method`/`--deployment` is
probably a heavily used option for most users, so I think it would
benefit from a shorthand option as it is quite verbose. Seems `-m` is
available and is kind of related.

I've also updated the docs with some references to this shorthand
version and replaced some longform usage with the short option, ensuring
it is made clear what the shorthand option is an alias for. Happy to
change this though.

### QC
<!-- Make sure that you can tick the boxes below. -->

* [ ] The PR contains a test case for the changes or the changes are
already covered by an existing test case.
* [x] The documentation (`docs/`) is updated to reflect the changes or
this is not necessary (e.g. if the change does neither modify the
language nor the behavior or functionalities of Snakemake).