test: Add test for issue #1256 and #2574 (#2583)
### Description

Add test for issues #1256 and #2574.

The test Snakefile has a syntax error on line 9. Some versions of
Snakemake, including 8.0.1, incorrectly report the error on line 20. I'm
only submitting the test case, not the fix.

### QC

* [x] The PR contains a test case for the changes or the changes are
already covered by an existing test case.
* [x] The documentation (`docs/`) is updated to reflect the changes or
this is not necessary (e.g. if the change does neither modify the
language nor the behavior or functionalities of Snakemake).