fix: bug with preemptible rules (#2616)
The second check in cli.py needs to be also looking at
args.preemptible_rules, and not the (not defined yet) variable that is
intended to be set.

### Description

<!--Add a description of your PR here-->

### QC
<!-- Make sure that you can tick the boxes below. -->

* [ ] The PR contains a test case for the changes or the changes are
already covered by an existing test case.
* [ ] The documentation (`docs/`) is updated to reflect the changes or
this is not necessary (e.g. if the change does neither modify the
language nor the behavior or functionalities of Snakemake).