feat: support for continuously updated input (using Python queues) (#2594)
This PR introduces the ability to define a Python queue as input. In
such a case, Snakemake continously extends the input files of a job
based on that queue and updates any implied dependencies in the DAG.
Once a finish sentinel is encountered, the job is considered final and
marked as ready for execution.

### QC
<!-- Make sure that you can tick the boxes below. -->

* [x] The PR contains a test case for the changes or the changes are
already covered by an existing test case.
* [x] The documentation (`docs/`) is updated to reflect the changes or
this is not necessary (e.g. if the change does neither modify the
language nor the behavior or functionalities of Snakemake).