fix: batch bug #2643 (#2650)
### Description

Fixes #2643. Seems to have been a refactor gone awry. Argparse "type"
fuctions (*a la* `add_argument(type=XXXX)`) take str and produce some
type. `parse_batch()` seems to expect an argparse namespace.

This wasn't caught in the unit tests, as they invoke `snakemake.run()`
directly in a way that bypasses the CLI arg parsing, which is where this
bug is. ~~Not sure what to do about that.~~ (wrote some tests)

### QC

* [x] The PR contains a test case for the changes or the changes are
already covered by an existing test case.
* [x] The documentation (`docs/`) is updated to reflect the changes or
this is not necessary (e.g. if the change does neither modify the
language nor the behavior or functionalities of Snakemake).