fix: Fix collect-lookup attribute error (#2687)
### Description

<!--Add a description of your PR here-->

This fixes #2685 with a simple test case added in `expand` function.

Please tell me if there are more elegant ways to fix this issue.

### QC
<!-- Make sure that you can tick the boxes below. -->

* [X] The PR contains a test case for the changes or the changes are
already covered by an existing test case. (No changes in documentation)
* [X] The documentation (`docs/`) is updated to reflect the changes or
this is not necessary (e.g. if the change does neither modify the
language nor the behavior or functionalities of Snakemake).

---------

Co-authored-by: Johannes Koester <johannes.koester@uni-due.de>