feat: allow for more extensive benchmark file in jsonl format (#2691)
### Description

<!--Add a description of your PR here-->
Another try at #2176, where we try to gather more information about a
job; for example, not only resources used, but also resources requested
and size of input files.
Here, the original `benchmark` file is left as is, but there is an
option to request an extended `benchmark` file if the extension is
`.jsonl`.

Is there any other relevant info to include?

@SichongP @corneliusroemer feel free to comment.

### QC
<!-- Make sure that you can tick the boxes below. -->

* [x] The PR contains a test case for the changes or the changes are
already covered by an existing test case.
* [x] The documentation (`docs/`) is updated to reflect the changes or
this is not necessary (e.g. if the change does neither modify the
language nor the behavior or functionalities of Snakemake).