fix: use keyword arguments in `_IOFile.open` (#2847)
fixes #2846

Previously, keword arguments to `.open` were not actually used leading
to unexpected (by me) behaviour. I don't think the documentation needs
to change to reflect this, because it's what one would expect in the
first place.

### QC
<!-- Make sure that you can tick the boxes below. -->

* [ ] The PR contains a test case for the changes or the changes are
already covered by an existing test case.
* [x] The documentation (`docs/`) is updated to reflect the changes or
this is not necessary (e.g. if the change does neither modify the
language nor the behavior or functionalities of Snakemake).

---------

Co-authored-by: Johannes Köster <johannes.koester@tu-dortmund.de>