feat: add flag to mark files where path should not be modified (#2888)
<!--Add a description of your PR here-->

When using modules with a prefix, the prefix is added to all input and
output files. If an input file is defined in the config, the path
becomes invalid. As such it is necessary to mark these files so that
their path are not changed.

Not sure if `path_modified` is the best name.... maybe `protect_path`,
or `not_modify_path`, or `no_prefix`, or `fixed_path`?

### QC
<!-- Make sure that you can tick the boxes below. -->

* [x] The PR contains a test case for the changes or the changes are
already covered by an existing test case.
* [x] The documentation (`docs/`) is updated to reflect the changes or
this is not necessary (e.g. if the change does neither modify the
language nor the behavior or functionalities of Snakemake).