docs: update FAQ - recommend ensure function for failing on empty output (#2910)
<!--Add a description of your PR here-->
Update the FAQ entry 'How do I make my rule fail if an output file is
empty?' to point the user to the `ensure(... , non_empty=True)` function
documentation.

Please check that the :ref: link (intended to point to
https://snakemake.readthedocs.io/en/stable/snakefiles/rules.html#ensuring-output-file-properties-like-non-emptyness-or-checksum-compliance)
is correct - I'm not familiar enough with sphinx to be 100% confident.

### QC
<!-- Make sure that you can tick the boxes below. -->

* [x] The PR contains a test case for the changes or the changes are
already covered by an existing test case.
* [x] The documentation (`docs/`) is updated to reflect the changes or
this is not necessary (e.g. if the change does neither modify the
language nor the behavior or functionalities of Snakemake).