Add `can_transfer_local_files` to executor plugin interface
Some plugins that don't assume a shared filesystem are capable of managing
their own file transfers from the workflow host to the remote execution points.
An example of this is HTCondor, which rarely assumes a shared filesystem between
the access point and the execution point, and where HTCondor is responsible for
transferring the files.

The `can_transfer_local_files` flag indicates the plugin -- and not snakemake --
is responsible for managing file movement between these two locations.