fix: parse f-string as-is from source (#2930)
<!--Add a description of your PR here-->

This is another (and hopefully the last) fix for f-string change in
python3.12.

Now everthing inner f-string are ignored, and only care about where
f-string start and end, and read it from snakefile as-is.

will fix #2881 and fix #2657
related PR: #2864, #2649

### QC
<!-- Make sure that you can tick the boxes below. -->

* [x] The PR contains a test case for the changes or the changes are
already covered by an existing test case.
* [x] The documentation (`docs/`) is updated to reflect the changes or
this is not necessary (e.g. if the change does neither modify the
language nor the behavior or functionalities of Snakemake).