fix: include more context to syntax errors when rule parsing fails (#2924) (#2926)
Updated three errors in rule parsing to throw `RuleException` instead of
`SyntaxError` - that way the user feedback includes line numbers to help
them identify the source of the error. Fixes issue #2924

### QC
* [ ] The PR contains a test case for the changes or the changes are
already covered by an existing test case.
* [x] The documentation (`docs/`) is updated to reflect the changes or
this is not necessary (e.g. if the change does neither modify the
language nor the behavior or functionalities of Snakemake).