feat: add `can_transfer_local_files` to executor plugin interface (#2921)
<!--Add a description of your PR here-->
Some plugins that don't assume a shared filesystem are capable of
managing their own file transfers from the workflow host to the remote
execution points. An example of this is HTCondor, which rarely assumes a
shared filesystem between the access point and the execution point, and
where HTCondor is responsible for transferring the files.

The `can_transfer_local_files` flag indicates the plugin -- and not
snakemake -- is responsible for managing file movement between these two
locations.

### QC
<!-- Make sure that you can tick the boxes below. -->

* [ ] The PR contains a test case for the changes or the changes are
already covered by an existing test case.
* [x] The documentation (`docs/`) is updated to reflect the changes or
this is not necessary (e.g. if the change does neither modify the
language nor the behavior or functionalities of Snakemake).

---------

Co-authored-by: JUSTIN HIEMSTRA <justin.hiemstra@path-ap2001.chtc.wisc.edu>
Co-authored-by: Johannes Köster <johannes.koester@tu-dortmund.de>
Co-authored-by: Johannes Koester <johannes.koester@uni-due.de>