chore: upgrade actions/checkout to v4 (#2946)
some of the CI actions still refer to actions/checkout@v3, which uses
the deprecated node.js v16 and which will soon stop working.

### QC
<!-- Make sure that you can tick the boxes below. -->

* [x] The PR contains a test case for the changes or the changes are
already covered by an existing test case.
* [x] The documentation (`docs/`) is updated to reflect the changes or
this is not necessary (e.g. if the change does neither modify the
language nor the behavior or functionalities of Snakemake).