feat: added snakemake.script.snakemake for type hinting (#2917)
The current magic injection of the `snakemake` object in python scripts
prevents linters and type checkers from recognizing that object. To
allow this the `snakemake.script` module is updated to include a
`snakemake` variable with only a type annotation of
`snakemake.script.Snakemake`, which has added type hints as well.
The preamble of a python script now also includes the following code
(here expanded over separate lines for readability):
```python
from snakemake import script
script.snakemake = snakemake # At this point the snakemake object is already unpickled
del script # This is not really necessary, but I think users should not expect this reference to be present
```

Existing scripts don't have to change, but should you want type checking
you can add
```python
from snakemake.script import snakemake
```
in your script, which then makes the object explicitly available in your
code.

There is no cross-talking between scripts because each script patches
its own copy of the `snakemake.script` module in memory as they all run
in separate interpreters.
Load time might be slightly increased, but when unpickling the snakemake
object the `snakemake.script` module is already imported. Memory usage
is also hardly changed as it is only an extra reference to the
`snakemake` object that needs to be stored.

### QC
<!-- Make sure that you can tick the boxes below. -->

* [x] The PR contains a test case for the changes or the changes are
already covered by an existing test case.
* [x] The documentation (`docs/`) is updated to reflect the changes or
this is not necessary (e.g. if the change does neither modify the
language nor the behavior or functionalities of Snakemake).