fix: unawaited coroutine sanitize_local_storage_copies (#2972)
Coroutine sanitize_local_storage_copies is unawaited, producing the
following warning:

```
<path>/lib/python3.12/site-packages/snakemake/dag.py:1615: RuntimeWarning: coroutine 'DAG.sanitize_local_storage_copies' was never awaited
  self.sanitize_local_storage_copies()
RuntimeWarning: Enable tracemalloc to get the object allocation traceback
```

This PR fixes it. 

### QC
<!-- Make sure that you can tick the boxes below. -->

* [ ] The PR contains a test case for the changes or the changes are
already covered by an existing test case.
* [ ] The documentation (`docs/`) is updated to reflect the changes or
this is not necessary (e.g. if the change does neither modify the
language nor the behavior or functionalities of Snakemake).

Co-authored-by: Anfeng Li <anfeng.li@cern.ch>
Co-authored-by: Johannes Köster <johannes.koester@tu-dortmund.de>