fix: allowing trailing '+' in name patterns (#3002)
This PR should fix issue #3001: 

A workflow might contain simple regexes within Python code, e.g.
`r"\s+"`. Prior to this PR the `NAME_PATTERN` of the linter considered
this first part, the `s` as part of a string to be concatenated with a
path.

### QC

There is no additional test (yet. Not sure it needs one). Docs do not
need updating as the intended behaviour does not change.

* [ ] The PR contains a test case for the changes or the changes are
already covered by an existing test case.
* [ ] The documentation (`docs/`) is updated to reflect the changes or
this is not necessary (e.g. if the change does neither modify the
language nor the behavior or functionalities of Snakemake).


<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->

## Summary by CodeRabbit

- **New Features**
- Improved name validation logic by updating the regex pattern, ensuring
that file names do not end with a '+' character, reducing the risk of
unintended exclusions.

- **Documentation**
- Added comments to clarify the implications of the regex changes and to
suggest alternatives for different separator needs.

<!-- end of auto-generated comment: release notes by coderabbit.ai -->