docs: Update installation.rst to recommend Miniforge instead of Mambaforge (#2975)
We should recommend people to use Miniforge, since it is now merged with
Mambaforge and from the README of the link

> Given its wide usage, there are no plans to deprecate Mambaforge. If
at some point we decide to deprecate Mambaforge, it will be
appropriately announced and communicated with sufficient time in
advance.
>
> That said, if you had to start using one today, we recommend to stick
to Miniforge.

In any case the link doesn't quite work since the `#mambaforge` part
doesn't actually bring you to the relevant part of the README

<!--Add a description of your PR here-->

### QC
<!-- Make sure that you can tick the boxes below. -->

* [x] The PR contains a test case for the changes or the changes are
already covered by an existing test case.
* [x] The documentation (`docs/`) is updated to reflect the changes or
this is not necessary (e.g. if the change does neither modify the
language nor the behavior or functionalities of Snakemake).