fix: Allow hyphens in config keys given on the command line. (#2998)
<!--Add a description of your PR here-->
It is possible to use hyphens when specifying config keys given in the
configfile, but when given on the command line as `--config
key-with-hyphens=value`, it fails. With this PR, hyphens are allowed in
the middle of config keys, when given on the command line.

As far as I understand, hyphens are valid for keys in the yaml
specification: https://yaml.org/spec/1.2.2/
https://www.yaml.info/learn/bestpractices.html

I have been using this code in my projects for a while and I don't think
it introduces any problems, especially since hyphens have been allowed
in the config file for a very long time.

The motivation for me to use hyphens in the command line config is
discussed here: https://github.com/snakemake/snakemake/issues/2986

Best, Carl

### QC
<!-- Make sure that you can tick the boxes below. -->

* [x] The PR contains a test case for the changes or the changes are
already covered by an existing test case.
* [x] The documentation (`docs/`) is updated to reflect the changes or
this is not necessary (e.g. if the change does neither modify the
language nor the behavior or functionalities of Snakemake).

Co-authored-by: Johannes Köster <johannes.koester@tu-dortmund.de>