fix: add assets and use local file links to allow offline reports (#2904)
This PR solves the issue with the offline reports. Several files are
downloaded on the fly while generating the report(js, css and licenses).
They are now added to the package as assets.

I used the `test_report` test case to verify that the generated report
is the same and works as expected. One minor change in the report
appears, as I thought it is more logical to use major.minor version
numbering for the `react` package as well.

To my opinion a update of the documentation is not necessary.

Closes #282 

### QC
<!-- Make sure that you can tick the boxes below. -->

* [x] The PR contains a test case for the changes or the changes are
already covered by an existing test case.
* [x] The documentation (`docs/`) is updated to reflect the changes or
this is not necessary (e.g. if the change does neither modify the
language nor the behavior or functionalities of Snakemake).


<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->
## Summary by CodeRabbit


- **New Features**
- Improved installation process with the addition of
`setuptools-download` for enhanced dependency management.
- New section added to manage the downloading of essential resources
during installation.

- **Bug Fixes**
- Enhanced file path handling to support both absolute and relative
paths, ensuring better file accessibility.

- **Chores**
- Reaffirmed inclusion of the `LICENSE.md` file in the package
distribution for compliance.

- **Refactor**
- Updated package resource management to utilize local paths, improving
reliability and portability.

<!-- end of auto-generated comment: release notes by coderabbit.ai -->

---------

Co-authored-by: Johannes Köster <johannes.koester@tu-dortmund.de>
Co-authored-by: Johannes Köster <johannes.koester@uni-due.de>
Co-authored-by: coderabbitai[bot] <136622811+coderabbitai[bot]@users.noreply.github.com>