docs: various documentation fixes (#3052)
<!--Add a description of your PR here-->

I initially wanted to fix only the Mambaforge mentions in the Setup
tutorial, but then noticed several warnings when building the docs so I
tried tackling those as well. Here is a list:

- Replace Mambaforge by Miniforge in the Setup tutorial
- Remove deprecated `sphinxawesome_theme.highlighting` extension
- Fix broken links in the docs (internal links and links to API docs)
- Fix headings that the did not show up in HTML (Bash examples)
- Fix warnings about duplicate external links (e.g. links named "here")
by making them anonymous with a double underscore.
- Replace sub-worfklows by modules in Additional Features tutorial about
modularization

Let me know if these are too unrelated for a single PR and you want me
to split it.

Thank you!

### QC
<!-- Make sure that you can tick the boxes below. -->

* [x] The PR contains a test case for the changes or the changes are
already covered by an existing test case. (N.A., I think)
* [x] The documentation (`docs/`) is updated to reflect the changes or
this is not necessary (e.g. if the change does neither modify the
language nor the behavior or functionalities of Snakemake).


<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->

## Summary by CodeRabbit

- **Documentation Updates**
- Removed outdated highlighting extension from documentation
configuration.
	- Updated references in cloud execution documentation for clarity.
- Enhanced hyperlink formatting throughout various documentation files
for consistency.
- Clarified terminology related to modularization and workflow
management.
- Revised tutorial content to reflect the transition from "Mambaforge"
to "Miniforge" for Python installation instructions.

<!-- end of auto-generated comment: release notes by coderabbit.ai -->