fix: Wrong linenumbers reported when linting (#2985)
This PR addresses two issues that led to incorrect linenumbers reported
when running `snakemake --lint`.

1. If an error was encountered in `--lint` mode, the stacktrace wasn't
cut to remove Snakemake internals and the linenumbers were not adjusted,
because the linemap was not available. Fixed by storing the workflow
object to make the linemap accessible.
2. For rule lints, the linemap wasn't applied, resulting in wrong
linenumbers in the output of `snakemake --lint`. Fixed by applying the
linemap before outputting the lints.

I added tests for both cases. 

Fixes #2574. 

### QC
<!-- Make sure that you can tick the boxes below. -->

* [X] The PR contains a test case for the changes or the changes are
already covered by an existing test case.
* [X] The documentation (`docs/`) is updated to reflect the changes or
this is not necessary (e.g. if the change does neither modify the
language nor the behavior or functionalities of Snakemake).


<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->
## Summary by CodeRabbit


- **New Features**
- Enhanced linting functionality with improved line number retrieval for
rules.
- Introduced a new method to centralize line number logic in the linting
process.
- Added a new test for handling specific error conditions in Snakemake
workflows.

- **Bug Fixes**
- Improved accuracy in asserting reported line numbers for unused
parameters during linting tests.

- **Documentation**
- New configuration file for testing, improving organization and
readability of test setups.

<!-- end of auto-generated comment: release notes by coderabbit.ai -->

---------

Co-authored-by: Johannes Köster <johannes.koester@tu-dortmund.de>